Add files via upload
Run the program. Draw a digit 0-9. Press enter and it'll try to guess your digit. Right click to clear the screen and you can try it again.

This is an algorithmic approach to handwriting recognition; not a NN or ML approach. There is also a confidence parameter so that you can exclude non-confident guesses.

The way it works is the program converts what you wrote to a series of points. The large number of points are then decimated to 20 points. It is then scaled up to a 1:1 aspect ratio to normalize the points so that you can draw the digits in any size or location. The set of points is then compared to 10 reference samples of the digits 0-9. Each point distance to the reference point is added up, and at the end, the one with the lowest overall distance is deemed the digit.